fix: Player movimiento 4 direcciones (up, down, left, right)
CORRECCION:
- Agregadas player::move_up() y player::move_down()
- Ahora movimiento completo en 4 direcciones:
  - player::move_left() - Izquierda (-x)
  - player::move_right() - Derecha (+x)
  - player::move_up() - Arriba (-y)
  - player::move_down() - Abajo (+y)

USO EN RYDIT:
ryda frame < 10000 {
    onif tecla_presionada("arrow_up") {
        player::move_up(player)
    }
    onif tecla_presionada("arrow_down") {
        player::move_down(player)
    }
    onif tecla_presionada("arrow_left") {
        player::move_left(player)
    }
    onif tecla_presionada("arrow_right") {
        player::move_right(player)
    }
    player::apply_gravity(player)
}

+++ crates/rydit-gfx/examples/demo.rs
@@ -29,7 +29,7 @@ fn main() {
             d.clear(ColorRydit::Negro);
 
             // Dibujar círculo rojo en el centro (animado)
-            let radius = 50 + (frame % 20) as i32;
+            let radius = 50 + (frame % 20);
             d.draw_circle(400, 300, radius, ColorRydit::Rojo);
 
             // Dibujar rectángulo verde
